refactor(session-broker): move the network decisions to network.rs
The refusal decisions sat in main.rs, which the coverage gate ignores
and which carries no dead_code allowance for a non-Linux host.

network.rs owns them now, with their tests. The argv test is the one
place that pins -q -n -t 7 -T 3 -s, and its name says how long the
seven attempts wait.
