+++ scripts/coverage-floor.sh
@@ -38,7 +38,7 @@ crates/lns-service/src/vm/cloud_hypervisor/process.rs    process syscall leaf (s
 crates/lns-service/src/vm/cloud_hypervisor/real.rs       VMM/vsock syscall leaf: CloudHypervisor::run (spawn_blocking/current-thread-runtime adapter into run_async_with) + the guest-vsock accept loop (UnixListener::accept) and into_blocking_fd (UnixStream into_std/into_raw_fd) whose OS-error arms can't be deterministically driven. run_async_with's orchestration + the connector/handshake logic are host-tested at 100% in cloud_hypervisor.rs and vsock.rs.
 crates/lns-service/src/approval_flow/system_font.rs      macOS Core Text leaf (CTFontCreateUIFontForLanguage → kCTFontURLAttribute → CFURL → file read) whose ? error arms can't be deterministically triggered; the pure font-definition building (build_font_defs/apply_host_fonts) is host-tested at 100% in window.rs, and resolution is proven on a real macOS host by system_font's own #[cfg(target_os="macos")] tests.
 crates/lns-cli/src/raw_mode/real.rs                      RealTty libc leaf (isatty/tcgetattr/tcsetattr/cfmakeraw/ioctl(TIOCGWINSZ)) + real_tty() wiring + the no-arg production entry points; Tty trait, *_with helpers and RawModeGuard Drop are host-tested at 100% via FakeTty in raw_mode.rs. Pinned end-to-end by the interactive-shell.exp smoke (manual)
-crates/lns-session-broker/src/main.rs                    guest broker entry; static-musl PID-1 sibling, only runs inside the microVM. Exit-code clamping lives in exit.rs and is host-tested at 100%.
+crates/lns-session-broker/src/main.rs                    guest broker entry; static-musl PID-1 sibling, only runs inside the microVM. Exit-code clamping lives in exit.rs and the network refusal decisions live in network.rs, both host-tested at 100%.
 crates/lns-session-broker/src/pty/real.rs               RealPtySyscalls libc leaf (posix_openpt/grantpt/unlockpt/ptsname/open/setsid/ioctl/getpgrp/tcsetpgrp/tcgetpgrp/dup2) + open_pty/child_setup/foreground_pgrp wiring; allocation+cleanup ordering, error mapping, dup2 loop and pgrp predicate host-tested at 100% in pty.rs via FakePty
 crates/lns-session-broker/src/session/real.rs           guest session driver leaf: fork + post-fork child arms + OS-thread pump/drive loops + libc (pty/pipe/signal/exec). The pure decisions (open-session parse, client-frame dispatch, signal targeting, argv validation) and SharedFd / read_client_frame are host-tested at 100% in session.rs.
 crates/lns-session-broker/src/vsock/real.rs             RealVsockSyscalls libc leaf (AF_VSOCK socket/bind/listen/accept/read/write) + listen/accept/read_exact/write_all wiring; the listen close-on-failure ordering, accept EINTR loop, sockaddr_vm construction, and read/write partial+EINTR+EOF loops are host-tested at 100% in vsock.rs via FakeVsock
